refactor: extract watchlist creation functionality into a separate WatchlistMenu component
