+++ .env
@@ -1,3 +1,3 @@
-VITE_API_HOST=http://stockserver-env-1.eba-5dyjx7cw.eu-north-1.elasticbeanstalk.com
+VITE_API_HOST=https://api.symbata.com
 #VITE_API_HOST=http://localhost:3000
